+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - fix random crashes on quote reply #4337
 - avoid drafts in readonly chats #4349
 - fullscreen images getting cropped a little #4402, #4385
+- settings: chat background preview element not working for the default background image #4403
 - macOS: make area under traffic lights dragable and fix the bug that its size changed based on profile acount and window height #4408
 - fix chat "scrolls up" right after switching #4404